+++ src/reanimated1/Easing.js
