+++ src/reanimated1/animations/Animation.js
