+++ src/reanimated1/animations/decay.js
