+++ src/reanimated1/animations/spring.js
