+++ src/reanimated1/animations/timing.js
