+++ src/reanimated1/core/AnimatedBezier.js
