+++ src/reanimated1/core/AnimatedBlock.js
