+++ src/reanimated1/core/AnimatedCall.js
@@ -1,5 +1,5 @@
 import invariant from 'fbjs/lib/invariant';
-import ReanimatedEventEmitter from '../ReanimatedEventEmitter';
+import ReanimatedEventEmitter from '../../ReanimatedEventEmitter';
 import { val } from '../val';
 import AnimatedNode from './AnimatedNode';
