+++ src/reanimated1/core/AnimatedClock.js
