+++ src/reanimated1/core/AnimatedClockTest.js
