+++ src/reanimated1/core/AnimatedFunction.js
