+++ src/reanimated1/core/AnimatedOperator.js
