+++ src/reanimated1/core/AnimatedProps.js
