+++ src/reanimated1/core/AnimatedSet.js
