+++ src/reanimated1/core/AnimatedStartClock.js
