+++ src/reanimated1/core/AnimatedStopClock.js
