+++ src/reanimated1/core/AnimatedStyle.js
