+++ src/reanimated1/core/Core.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import Animated from '../Animated';
+import Animated from '../../Animated';
 
 import renderer from 'react-test-renderer';
 
-jest.mock('../ReanimatedEventEmitter');
-jest.mock('../ReanimatedModule');
-jest.mock('../reanimated2/NativeReanimated');
+jest.mock('../../ReanimatedEventEmitter');
+jest.mock('../../ReanimatedModule');
+jest.mock('../../reanimated2/NativeReanimated');
 
 describe('Core Animated components', () => {
   xit('fails if something other then a node or function that returns a node is passed to Animated.Code exec prop', () => {
