+++ src/reanimated1/core/__mocks__/AnimatedProps.js
