+++ src/reanimated1/core/createEventObjectProxyPolyfill.js
