+++ src/reanimated1/derived/__mocks__/evaluateOnce.js
