+++ src/reanimated1/derived/color.js
