+++ src/reanimated1/derived/diff.js
