+++ src/reanimated1/derived/diffClamp.js
