+++ src/reanimated1/derived/evaluateOnce.js
