+++ src/reanimated1/derived/interpolateColors.js
