+++ src/reanimated1/operators.js
